fix(shred-collector): turbine tree node list wrong capacity
the list of all nodes should be large enough to fit all staked nodes + gossip nodes + 1 for self. the bug was that it wasn't including 1 for itself, and it would hit an error on appendAssumeCapacity.

normally this is not a problem because nodes get filtered out in the for loop, so it doesn't use the full capacity. but if no nodes are filtered out, then you get an error from having a capacity 1 less than necessary.

this also refactors the stake calculation if-statement for clarity