feat(shred-network): minimize dependencies (#462)
This is an alternative implementation for #423. It accomplishes the same goals, except that it implements the interface between `shred-network` and an epoch context provider with shared memory instead of a vtable.

The static context of an epoch is defined by EpochContext, which contains staked nodes and a leader schedule.

The storage of relevant EpochContexts is managed by EpochContextManager. This specialized struct is needed because it must be aware of the epoch schedule.

Depending on the run mode, the initial EpochContext may be created from the bank or from rpc. There is currently only one way to update the EpochContextManager during the runtime of the validator, which is via rpc.

There is a limitation where RPC cannot be used to access epoch staked nodes. This can only be acquired from the bank. So the bank is used when available, otherwise the data is not populated.

## commits

* refactor(leader_schedule): SlotLeaderProvider from PointerClosure to direct SlotLeaders interface

* refactor(sync): adapt RcSlice for single-item pointers

* feat: add Window and SharedPointerWindow data structures

* feat(shred-network): minimize dependencies

* fix(collections): window integer overflow when realigning from 0. add test to reproduce issue and fix bug

* fix(sync): shared pointer window defer lock should be unlock

* fix(rpc epoch context): wrong epoch calculation