fix(ledger,shred-network): memory leak (#510)
* fix(shred-network): keep up with turbine

* feat(shred-network): simplify repair logic, remove metric-like logging, try to complete slots more eagerly

* fix(shred-network): tests  + linting

* fix(shred-network): test

* fix(repair) add jitter to handle edge case where all the following are true:

- we're caught at least 1024 slots behind or turbine isn't working
- the current bottom slot is skipped
- the next slot is skipped
- the slot 25 slots ahead is also skipped

with jitter, we only run into a problem if there are also 40 consecutive slots that are skipped, which is practically impossible

* grafan: update shred collector dashboard

* refactor(ledger): switch formatting

Signed-off-by: Drew Nutter <dnut@users.noreply.github.com>

* refactor(shred-network): function spacing

Signed-off-by: Drew Nutter <dnut@users.noreply.github.com>

* feat(time): expand Instant to support shred repair

Signed-off-by: Drew Nutter <dnut@users.noreply.github.com>

* const shred wip

* refactor(ledger): decouple shred mutations from shred struct

this is a first step towards shreds themselves being immutable. the mutations now act on a mutable shred slice instead of the shred itself. but the shred still needs to be mutated sometimes so payloadMut serves this purpose until mutations of Shred is eliminated

* refactor: clean up shred inserter params with options struct

* fix(ledger): shreds leak - clarify shred lifetimes and clean them up

previously any shreds created in shred processor or insertShreds leaked.

this establishes clarified ShredInserter.insertShreds lifetime rules:
- shreds passed in are owned by caller and must be freed by caller (applied this in shred processor)
- shreds created in the function need to be cleaned up in the function, unless...
- shreds returned from the function need to be owned by the caller, so they need a sufficient lifetime that exceeds the state of insertshreds and the shreds that were passed in. this is currently solved by duping the shreds.

* fix: memory errors and rocksdb writebatch leak

* fix(shred-network): reduce repair service peer cache size by factor of 16

in my test i saw its memory usage go down from 1 GB to 6 MB which is obviously a much bigger drop than 16x, so I'm not sure how easy it is to isolate the impact of this change. Nonetheless I don't think this cache needs to go for 128 slots since typically repair does not need to go back more than a few slots. there is some potential to optimize this for the startup process of catching up from behind. but it's working as is in my testing

* fix: style

* rename rng to prng

* fix timer usage

* refactor(time,repair): use Duration for repair sleep delay

* remove unused

* rename shred collector to shred network

* docs: shred network cli args

* docs(ledger): shred inserter lifetime rules

---------

Signed-off-by: Drew Nutter <dnut@users.noreply.github.com>