fix(gossip): leak and double free on overwritten items
when overwriting, there was a bug (recently introduced) where the overwritten item is supposed to be returned to the caller, and they deinit it. instead, the *new* item was being returned as if it was the overwritten item. the caller would deinit the new item, leaving corrupted memory in the gossip table (which would eventually be double freed). the overwritten item would never be freed, so it leaked

also adds a test and adds a stack trace to an error log