refactor(collections): simplify SplitUnionList to remove overarching usize index, just using tagged index
this simplificaiton means you can no longer index with an integer over all items. this is a limitation, but this extra indexing makes the data structure use a surprising amount of memory. instead i've opted to make the data structure a simpler building block, and the user of the struct can implement a wrapper with this indexing if needed.

in this case, its internal indexing was redundant with indexing that GossipMap is already doing. so by stripping out this indexing, GossipMap doesn't need any extra logic to handle it. i've just simplified the code and reduced the memory usage.