perf(gossip): reduce gossip table memory usage (#526)
* perf(gossip): reduce gossip table memory usage

* fix(gossip): leak and double free on overwritten items

when overwriting, there was a bug (recently introduced) where the overwritten item is supposed to be returned to the caller, and they deinit it. instead, the *new* item was being returned as if it was the overwritten item. the caller would deinit the new item, leaving corrupted memory in the gossip table (which would eventually be double freed). the overwritten item would never be freed, so it leaked

also adds a test and adds a stack trace to an error log

* feat(collections): implement SplitUnionList to store unions more efficiently

* refactor(gossip): use SplitUnionList in GossipMap instead of MultiArrayList to save space

* refactor(collections): simplify SplitUnionList to remove overarching usize index, just using tagged index

this simplificaiton means you can no longer index with an integer over all items. this is a limitation, but this extra indexing makes the data structure use a surprising amount of memory. instead i've opted to make the data structure a simpler building block, and the user of the struct can implement a wrapper with this indexing if needed.

in this case, its internal indexing was redundant with indexing that GossipMap is already doing. so by stripping out this indexing, GossipMap doesn't need any extra logic to handle it. i've just simplified the code and reduced the memory usage.

* fix: style

* fix: style

* docs(collections): note implicit assertion in SplitUnionList

* delete stale comment that is now inaccurate

* document SplitUnionList

* remove commented code from test

* clean up unnecessary inlining of literal enums

* docs formatting

* clean up EnumStruct issues that were copied from std

* move gossip map init in gossip table to init function

* use iterator in test

* document GossipMap

* extract Entry from GossipMap