feat(ci): run and update service (#579)
This adds some files that can be installed into a linux system to run
sig as a long running service, and periodically have it self-update
based on the latest code and restart.

This also adds an option to the logger to output to a file, and makes
stderr logging optional.

See the
[readme](https://github.com/Syndica/sig/blob/dnut/ci/runupdate/ci/run-and-update-service/README.md)
for more info.

I added a folder in the repo root called "ci" for this since no existing
folders seemed appropriate.