refactor(cmd): mv src/cmd/config.zig to src/config.zig (#582)
I find it confusing that we have a cmd.zig in the sources root, and then
we have a separate cmd folder that has nothing except the application
configuration. I don't see why this folder exists if it doesn't contain
cmd.zig. Either the folder should contain both cmd.zig and config.zig,
or the folder should be eliminated and both go in src. Dividing it in
half is just confusing and seems like the worst approach. In this PR
I've opted to eliminate the folder.