fix(prometheus): http server exit on error (#590)
Currently, if the prometheus http server has an error, it just prints
the name of the error to stderr and then exits the thread. Metrics stop
being served. All kinds of errors could happen in here, but that doesn't
mean we should stop serving metrics. It's better to log an error and
then continue running. I used runService from the service manager to
accomplish this, with a bit of refactoring in how the thread is spawned.