refactor(rpc-client): redesign RPC client (#593)
I redesigned the RPC client to make the code more organized and
composable, without adding unnecessary abstractions such as interfaces
or generic types. The RPC client struct itself is now simplified, and
only serves to glue together the code from other modules: `methods`,
`http`, `request`, and `response`.

## Motivation

My primary motivation for these changes was to improve the organization
and readability of the code. Each of these domains (networking,
serialization, solana data model) are disjoint and deal with a different
set of responsibilities. I feel the code is easier to understand when
they are clearly delineated in separate modules.

An extra benefit of this organization is that the RPC client is more
composable and flexible. You can easily combine these components, or
provide custom implementations for them, to implement whatever kind of
RPC client you would like in a few lines of code. The provided RPC
client glues together these pieces for ease of use.

A common use case that I've needed in the past when I was writing
trading bots was to blast out many rpc requests and then collect the
responses asynchronously. This required some ugly hacks using the rust
rpc client. With my new approach here, use cases like that are much
easier to support through composition of a custom http client.

## Approach

RPC Methods, from the perspective of an RPC client, are just data. They
are request and response data types that need to be converted back and
forth from json. So I've described that fully with the type system. Now
there is `methods.zig` with a struct for each method. These structs can
be used in the client and server.

I also extracted serialization and HTTP connectivity into their own
modules. Serialization is implemented in `request.zig` and
`response.zig`. The HTTP logic was moved into `HttpPostFetcher` which is
defined in `http.zig`.

Thus in client.zig we have the full implementation of an rpc client in
the `fetchRpc` function. It runs these steps in 5 lines of code:
1. serialize the request as json
2. send json in an http post message and await the response
3. deserialize the json as the response type

While `fetchRpc` sufficiently implements an rpc client, a more
opinionated `Client` struct is also provided to improve its ergonomics.
It handles the following:
- initializes and contains an HttpPostFetcher
- defines two fetch methods: one to use the contained allocator, and one
to use a custom allocator
- defines a helper method for each RPC method so you can call them like
`client.getSlot(.{})`

## Other improvements

### Request
I was able to simplify the code by removing the `Request` data type. It
is no longer necessary since request structs can be directly serialized.

### Response

I reorganized the Response data structure to be more intuitive and safe.
It no longer has intermediary states where the struct is partially
invalid. You just init it from the json to get a struct that's more
neatly organized than the old version.

### Tests

I removed the dependency on testnet from the rpc tests. The existing
tests were validating three things:
1. serialization and deserialization works
5. we're able to connect to a testnet rpc node with http requests
6. the testnet rpc node associates our request data type with the same
response data type that we do.

I converted the tests to purely validate `1`. Validations `2` and `3`
are no longer being tested because it's not appropriate to test these in
a unit test. This is why the tests previously were skipped. So I've
actually increased coverage by unskipping the tests. We could also add
tests similar to `2` that connect HttpPostFetcher to a *local* server,
but I think that's overkill, since the http client comes from `std` and
involving the network in unit tests is a headache and leads to flaky
tests.