fix(service-manager): truncate overlong thread names (#606)
Linux only allows thread names up to 15 characters. If you exceed that
limit, setting the thread name fails and the thread doesn't get a name.

One solution is to just always name services with names under 15
characters. But those names are used in log messages and it would be
nice to use more characters. So instead, I've just accepted that we're
going to use longer names, and the thread name used by the kernel needs
to be truncated to 15 characters.

I debated throwing a log message in here to warn about the name being
truncated. But I think it's just clutter in the logs since we probably
won't be trying to squeeze every name into 15 characters.