fix(adapter): reduce error log noise from RpcEpochContextService (#604)
I see this error pretty often in the logs:

```
time=2025-03-06T22:25:37.267Z level=error scope=adapter.RpcEpochContextService message="failed to refresh epoch context via rpc: error.EndOfStream"
```

This just means the connection was interrupted when downloading data
from an rpc node. It's not a serious error, and a single occurrence does
not imply any action needs to be taken by a human, so I added a retry
mechanism. If this error happens, we just ignore it and try again, for
up to three attempts. If all three fail, or any other error occurs, then
we log an error.