fix(build.zig): ssh step dependencies (#607)
Two bugs are fixed here:

`send-file.zig` requires libc due to using c_allocator.

Using options `ssh-host` and `no-run` means the binary should be sent to
the remote server but not run. It was only building locally and not
sending it. It would only send it when it's supposed to run (i.e. no-run
was not specified). This is because the build step dependencies were
incorrectly specified.