feat(svm): implement executable verification (#599)
The SVM runs a "verification" pass before executing the program. This
pass checks all of the instructions, not just the ones that happen to be
run. So, if, for example, some invalid code was hiding behind a branch
that was never reached, it will find it anyway.

This was a nice changeset, as it allowed me to remove the silly
`sbpf.Instruction.disallowed` map, which was just a hack around the true
verifier.

This PR hasn't been fuzzed yet, as I haven't set up the sig-fuzz
bindings yet, doing that next. We should be good to merge it anyway
before that, it has plenty of testing coverage, I will open follow-up
fixes later.