Add tests for lazy (#600)
Add some more tests for some utils modules. Including a test that was
missing from #596 (test a double peek).

While writing the unit-tests for Lazy, I corrected a small issue where
the final result was not actually returned. Most likely this was dormant
in the code because only `Lazy(void)` is currently used.

---------

Co-authored-by: David Rubin <87927264+Rexicon226@users.noreply.github.com>