gossip: improve docs + rearrange fields for counter-intuitive layout compatibility
zig throws the signature at the *end* of SignedGossipData which is a bit counterintuitive. so i flipped around VersionedGossipData to make it consistent. GossipData is always the first field, and the signature offset after that is the same - for both structs

i'm guessing this doesn't actually make a practical difference, but the order is arbitrary anyway, and in theory it might make the signedData method a tiny bit more efficient, since the value it needs to return is an exact copy of the leading bytes of the input struct