refactor(gossip): improve gossip map usage (#634)
This refactors the usages of the `GopssipMap` to improve memory access
patterns and readability. It also fixes a bug in `GossipDumpService`.

In most cases, we only need either the `Metadata` or the `GossipData`
from the map. And the map itself is already structured to segregate
these in memory. So it's strange that when the map is used, it typically
consolidates all the data together, only for the caller to extract out
only the single piece that is needed. The major theme of this PR is to
specialize these accesses to cater to both the usage and the internal
memory of the map.

See these commits for a description of the specific changes introduced
by this PR:
- f29f8ffd5df2947c112f4ab6d6a768b8bd0a43e6 fix(gossip): dump service
encoding buf is too small
- b08f26b9c31ea3666b70bfc7df84206652d64f9b refactor(gossip): remove
pointless GossipMap.KeyToIndex alias
- b35d164690452902309727d378d6aa19f33ab16f perf(gossip): avoid using
gossip map iterator when we only need metadata
- f6c2b6a02b66cb662b68c27649dfde25b725510b refactor(gossip): reorganize
GossipVersionedData into data+metadata and propagate this separation
throughout sig. usually we only need one or the other, so access them
independently when possible. this minimizes dependencies, makes the code
clearer, and should help performance too