feat(svm): implement data direct mapping (#679)
For now, the sBPF tests haven't moved over yet, since I ran into some
trouble with the Zig SDK. There are a lot of tests for the
`MemoryChunkIterator`, but we'll still need some for interacting with
the accounts themselves. There's the `direct_mapping` test, which
showcases a basic account data mutation.