feat(vm): syscall conformance (#682)
**ExecutionError**
The main change in this PR is the definition of the ExecutionError which
is an error set consisting of the union of InstructionError, EbpfError,
SyscallError, and some Sig specific errors.
InstructionError.ArithmeticOverflow has been renamed to
InstructionError.ProgramArithmeticOverflow to avoid conflicting with
SyscallError.ArithmeticOverflow. It is necessary to understand which
error set an error belongs to for conformance and logging.

---------

Co-authored-by: David Rubin <david@vortan.dev>